game: add ability to click on numbers
Just like the game on windows, it will uncover neighboring cells, if the
count of neighboring flags matches that number.
Added a test for that as well.
Allocate more events since now we log more of them, including invalid
player inputs - so now we can overflow if we spam them.