game: rework number discover
In some cases the number discover might trigger second hand invalid
triggers because uncover() is called inside.
We now remove the call and handle the uncover code directly.
Discovers are also recorded the same way as the 'discover many' event.
Check in the tests that a number discover actually discovers cells.